Refactor the build process to use package.json
Updates the build process to use package.json scripts for a much simpler
process.

Also, ship dist with git repo (for now at least)

Because I'm working on deployment stuff, I need to be able to reference
Denali by git repo to avoid publishing a couple dozen patch versions as
I iterate through small bugs. But because it's the git repo, it's
missing the dist folder (which is built and ships with the npm module,
or happens to exist when locally linked).

So for now, going to include dist in the repo.