Fix mixin inheritance and application
The old approach of manually recreating the subclassing via
setPrototypeOf had two problems:

1. setPrototypeOf has serious and far-reaching perf impacts
2. Intermediate classes wouldn't call super for omitted methods. If a
subclass called `super()`, and the intermediate class didn't define that
method, the super failed, which doesn't comply with ES2015 class spec.

The solution is to require the mixin factories to subclass the provided
base class themselves. It push a little more awareness onto the
consumer, which isn't great, but it's the only viable, robust option.