Refactor Model.create(), remove ORMAdapter.createRecord()
The proxied nature of models means that bulk assignment in the ORM layer
of attributes to a new record could result in properties intended for
the proxy being dropped (i.e. a `password` virtual property that a User
model uses to create `hashedPassword` on save).

So Model.create() builds an empty record and then assigns the given
attributes at the proxy layer, allowing the proxy to determine which
properties should fall through to the record.