drop top-level self links by default for JSONAPI
The incoming URL isn't the right source for determining the top-level
self link (a `POST /books` that creates a new book shouldn't return
`/books` as the self link for the newly created book). It's the wrong
abstraction here.

To do this right, we need semantic knowledge of the routes, i.e.
`router.resource()`. That kind of hook doesn't exist right now, so
punting.