fix(cli): run processEach hooks _after_ processParent hooks
processParent hooks will naturally follow the correct addon load order
thanks to the sorting that `discoverAddons()` provides.

However, the `before` and `after` options that inform that sort order
for addons via the discoverAddons method cannot fully specify load order
for nested structures - only siblings.

In other words, if I have the following addon graph:

    App
      Addon Alpha
        Addon Charlie
      Addon Bravo
        Addon Dog

Addon Dog is unable to specify it's load order relative to Addon
Charlie. The App may want Bravo to load _before_ Alpha, but Dog wants to
load _after_ Charlie

We could support this by expanding the load order configuration API to
support more complex constraints, but the likelihood of needing such
constraints is low, and the effort to implement high.

So for now, we'll assume that processEach hooks generally tend to run
after processParent hooks. We make this assumptiong because processEach
hooks must be extremely conservative in their assumptions, given that
they operate on an implied set of addons that have not explicitly opted
in to their effects. This conservatism means it's more likely that the
hook can operate on the compiled app vs. the source, since the compiled
version must match Denali conventions, but the source could deviate
wildly.

Meanwhile, processParent hooks can make more assumptions because the
user must explicitly opt in to their effects. Plus, load ordering
applies to the processParent hooks, so users can control their
execution.