fix: remove unused babel-register dep, convert denali-build.js blueprints to untranspiled syntax
While it is theoretically possible to transpile the denali-build.js
file, it's difficult to do by default (because Broccoli cannot pick a
file from the root directory). Thus, you'd have to create a "fake" tree
that doesn't read in from it's input node, but rather reaches out of
`tmp/` to grab the denali-build.js file directly and transpile it. Since
that gets complicated, we're favoring the simpler option for now, and
just defaulting denali-build.js files to standard, untranspiled syntax.