fix(cli): reinstate babel-register
The reason for this ignore method is complex.

Ideally, we want to avoid babel-register recompiling already transpiled
files - for perf reasons, and because of babel/babel#4946.  When the
denali CLI runs, it will attempt to load the local Denali's transpiled
files (i.e.  commands/index.js), but those are loaded from the
node_modules folder, which aren't transpiled by default, so that's okay.

The problem is when we are symlinking denali into the dependencies.
Babel resolves the realpath of those symlinks, so it doesn't recognize
that they are inside node_modules, and therefore won't match the default
node_modules ignore pattern. So we need to manually omit denali files
when they are symlinked in; hence the `denali/dist` ignore pattern,
which should only match for the denali source itself.

A general form of this problem exists, where a user attempts to import
from any Denali addon that is symlinked from the user's denali-build.js.
Babel will have similar difficulty recognizing the symlinked file as a
node_modules file, and so will attempt to transpile it.

However, that seems unlikely enough for now. The user has to import from
a denali addon, in their denali-build.js, the addon must be symlinked,
and the imported file must trip the above linked bug to trigger any user
facing error.