refactor(data): remove proxies from model class
Leverage the new onLoad hook supplied by the container to dynamically
add methods to the class definition for the various attribute and
relationships the user defines.

This drastically simplifies the Model class implementation, and ensures
that users will work with the Denali API rather than comingling the
ORM interface with the Denali interface.

fixes #338