fix(data): allow models to overwrite previous augmentations
This is slightly hacky, but basically the model augmentation approach we take
in onLoad slightly violates the container isolation, since it modifies the
model prototype which is shared across containers.

However, these property descriptors are based on the static model definition,
so therefore _should_ be identical across containers as well, so overwriting
isn't a problem.

It's possible that in a test, you could have static properties change from
test to test, which would introduce problems as those changes persist
between tests, but the proper way to do that would be to subclass rather
than directly mutate the class under test. That new subclass would be scoped
to that test's container, which should avoid problems.