feat: require container as arg to any DenaliObject
This is mostly a DX feature. The goal is to throw when the user
tries to directly instantiate a class that was intended for container
access. A good example of when this happens is in unit tests - you
cannot simply `import` and `new Foo()` if Foo extends DenaliObject
Of course you could just pass an empty object if you do not actually
_need_ the container for your unit tests, but it at least gets
you thinking, esp. for `init()` gotchas